# docs/apifox-full.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AuraStack 接口文档（自动生成）",
    "version": "1.0.0",
    "description": "AuraStack 接口文档（由脚本自动生成）。"
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Development"
    },
    {
      "url": "http://localhost:5000",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Flask Session Cookie 登录态"
      }
    }
  },
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "后台-用户管理"
        ],
        "summary": "查询users",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/用户管理",
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "后台-用户管理"
        ],
        "summary": "创建users",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          },
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "x-apifox-folder": "后台/系统管理/用户管理",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/admin/users/{user_id}": {
      "delete": {
        "tags": [
          "后台-用户管理"
        ],
        "summary": "删除users",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/用户管理",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "路径参数：user_id"
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "后台-用户管理"
        ],
        "summary": "更新users",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/用户管理",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "路径参数：user_id"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/api/admin/users/export": {
      "post": {
        "tags": [
          "后台-用户管理"
        ],
        "summary": "导出users",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/用户管理",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/admin/users/template": {
      "get": {
        "tags": [
          "后台-用户管理"
        ],
        "summary": "下载users导入模板",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/用户管理",
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/admin/users/import": {
      "post": {
        "tags": [
          "后台-用户管理"
        ],
        "summary": "导入users",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/用户管理",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
  }
}